Refactor builder
