+++ src/dpsnotificationservice/src/test/java/ca/bc/gov/open/pssg/rsbc/dps/dpsnotificationservice/OutputNotificationEndpointTest.java
@@ -1,7 +1,7 @@
 package ca.bc.gov.open.pssg.rsbc.dps.dpsnotificationservice;
 
 import ca.bc.gov.open.pssg.rsbc.dps.dpsnotificationservice.generated.models.*;
-import ca.bc.gov.open.pssg.rsbc.dps.files.notification.OutputNotificationMessage;
+import ca.bc.gov.open.pssg.rsbc.dps.notification.OutputNotificationMessage;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
