--- src/spd-notification-worker/src/main/java/ca/bc/gov/open/pssg/rsbc/spd/notification/worker/document/DpsDataIntoFigaroRequestBodyAdapterToMapper.java
@@ -1,8 +1,0 @@
-package ca.bc.gov.open.pssg.rsbc.spd.notification.worker.document;
-
-import org.mapstruct.Mapper;
-
-@Mapper
-public interface DpsDataIntoFigaroRequestBodyAdapterToMapper {
-    public DpsDataIntoFigaroRequestBody dpsDataIntoFigaroRequestBodyAdapterToMapper(DpsDataIntoFigaroRequestBodyAdapter dpsDataIntoFigaroRequestBodyAdapter);
-}
